#16 Cart

+++ server/controllers/blogCategoryController.js
@@ -5,7 +5,7 @@ const createCategory = asyncHandler(async (req, res) => {
     const { title } = req.body;
     const category = await BlogCategory.findOne({ title });
     if (category) {
-        throw new Error('This category already exists');
+        throw new Error('This blog category already exists');
     } else {
         const newCategory = await BlogCategory.create(req.body);
         return res.json({
@@ -15,29 +15,39 @@ const createCategory = asyncHandler(async (req, res) => {
     }
 });
 
-const getCategories = asyncHandler(async (req, res) => {
-    const response = await BlogCategory.find().select('title _id');
-    return res.json({
-        success: response ? true : false,
-        blogCategories: response ? response : 'Cannot get blog-categories'
-    });
-});
-
 const updateCategory = asyncHandler(async (req, res) => {
     const { bcid } = req.params;
-    const response = await BlogCategory.findByIdAndUpdate(bcid, req.body, { new: true });
-    return res.json({
-        success: response ? true : false,
-        updatedCategory: response ? response : 'Cannot update blog-category'
-    });
+    const { title } = req.body;
+    const category = await BlogCategory.findOne({ title });
+    if (category) {
+        throw new Error('This blog category already exists');
+    } else {
+        const response = await BlogCategory.findByIdAndUpdate(bcid, req.body, { new: true });
+        return res.json({
+            success: response ? true : false,
+            updatedCategory: response ? response : 'Cannot update blog category'
+        });
+    }
 });
 
 const deleteCategory = asyncHandler(async (req, res) => {
     const { bcid } = req.params;
-    const response = await BlogCategory.findByIdAndDelete(bcid);
+    if (!bcid) {
+        throw new Error('This category not exists');
+    } else {
+        const response = await BlogCategory.findByIdAndDelete(bcid);
+        return res.json({
+            success: response ? true : false,
+            deletedCategory: response ? response : 'Cannot delete blog category'
+        });
+    }
+});
+
+const getCategories = asyncHandler(async (req, res) => {
+    const response = await BlogCategory.find().select('title _id');
     return res.json({
         success: response ? true : false,
-        deletedCategory: response ? response : 'Cannot delete blog-category'
+        blogCategories: response ? response : 'Cannot get blog categories'
     });
 });
